refactor(host-application): split opencode runtime into focused modules
