+++ apps/desktop/src-tauri/crates/host-application/src/app_service/opencode_runtime/mcp_config.rs
@@ -1,0 +1,112 @@
+use anyhow::{anyhow, Context, Result};
+use host_infra_system::{command_exists, resolve_central_beads_dir};
+use serde_json::json;
+use std::path::Path;
+
+pub(crate) fn parse_mcp_command_json(raw: &str) -> Result<Vec<String>> {
+    let parsed: serde_json::Value =
+        serde_json::from_str(raw).context("Invalid OPENDUCKTOR_MCP_COMMAND_JSON format")?;
+    let values = parsed
+        .as_array()
+        .ok_or_else(|| anyhow!("OPENDUCKTOR_MCP_COMMAND_JSON must be a JSON string array"))?;
+
+    let command = values
+        .iter()
+        .map(|entry| {
+            entry
+                .as_str()
+                .map(str::trim)
+                .filter(|value| !value.is_empty())
+                .map(str::to_string)
+                .ok_or_else(|| {
+                    anyhow!("OPENDUCKTOR_MCP_COMMAND_JSON must contain only non-empty strings")
+                })
+        })
+        .collect::<Result<Vec<_>>>()?;
+
+    if command.is_empty() {
+        return Err(anyhow!("OPENDUCKTOR_MCP_COMMAND_JSON cannot be empty"));
+    }
+    Ok(command)
+}
+
+pub(crate) fn default_mcp_workspace_root() -> Result<String> {
+    let from_env = std::env::var("OPENDUCKTOR_WORKSPACE_ROOT")
+        .ok()
+        .map(|value| value.trim().to_string())
+        .filter(|value| !value.is_empty());
+    if let Some(root) = from_env {
+        return Ok(root);
+    }
+
+    let compiled_path = Path::new(env!("CARGO_MANIFEST_DIR"));
+    let root = compiled_path.ancestors().nth(5).ok_or_else(|| {
+        anyhow!("Unable to resolve OpenDucktor workspace root from manifest path")
+    })?;
+    Ok(root.to_string_lossy().to_string())
+}
+
+pub(crate) fn resolve_mcp_command() -> Result<Vec<String>> {
+    if let Ok(raw) = std::env::var("OPENDUCKTOR_MCP_COMMAND_JSON") {
+        return parse_mcp_command_json(raw.as_str());
+    }
+
+    if command_exists("openducktor-mcp") {
+        return Ok(vec!["openducktor-mcp".to_string()]);
+    }
+
+    if !command_exists("bun") {
+        return Err(anyhow!(
+            "Missing MCP runner. Install `openducktor-mcp` on PATH or install bun for workspace fallback."
+        ));
+    }
+
+    let workspace_root = default_mcp_workspace_root()?;
+    let direct_entrypoint = Path::new(&workspace_root)
+        .join("packages")
+        .join("openducktor-mcp")
+        .join("src")
+        .join("index.ts");
+
+    if direct_entrypoint.exists() {
+        return Ok(vec![
+            "bun".to_string(),
+            direct_entrypoint.to_string_lossy().to_string(),
+        ]);
+    }
+
+    Ok(vec![
+        "bun".to_string(),
+        "run".to_string(),
+        "--silent".to_string(),
+        "--cwd".to_string(),
+        workspace_root,
+        "--filter".to_string(),
+        "@openducktor/openducktor-mcp".to_string(),
+        "start".to_string(),
+    ])
+}
+
+pub(crate) fn build_opencode_config_content(
+    repo_path_for_mcp: &Path,
+    metadata_namespace: &str,
+) -> Result<String> {
+    let mcp_command = resolve_mcp_command()?;
+    let beads_dir = resolve_central_beads_dir(repo_path_for_mcp)?;
+    let config = json!({
+        "logLevel": "INFO",
+        "mcp": {
+            "openducktor": {
+                "type": "local",
+                "enabled": true,
+                "command": mcp_command,
+                "environment": {
+                    "ODT_REPO_PATH": repo_path_for_mcp.to_string_lossy().to_string(),
+                    "ODT_BEADS_DIR": beads_dir.to_string_lossy().to_string(),
+                    "ODT_METADATA_NAMESPACE": metadata_namespace,
+                }
+            }
+        }
+    });
+    serde_json::to_string(&config).context("Failed to serialize OpenCode MCP config")
+}
